+++ package.json
@@ -54,7 +54,7 @@
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
     "@types/jest": "^29.5.14",
-    "@types/node": "^22.13.4",
+    "@types/node": "^22",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "cors": "^2.8.5",
