+++ src/domains/post/list/_hooks/usePosts.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
-import { BASE_URL } from '@lib/api/constants';
+import { BASE_URL } from '@lib/api/url';
 import { Post } from '@lib/posts/postFactory';
 
 export type PostsResponse = {
